Move zone helpers to zone root
