+++ homeassistant/components/zone/zone.py
@@ -1,58 +1,11 @@
 """Zone entity and functionality."""
 from homeassistant.const import ATTR_HIDDEN, ATTR_LATITUDE, ATTR_LONGITUDE
 from homeassistant.helpers.entity import Entity
-from homeassistant.loader import bind_hass
-from homeassistant.util.async_ import run_callback_threadsafe
 from homeassistant.util.location import distance
 
-from .const import DOMAIN
-
-ATTR_PASSIVE = 'passive'
-ATTR_RADIUS = 'radius'
+from .const import ATTR_PASSIVE, ATTR_RADIUS
 
 STATE = 'zoning'
-
-
-@bind_hass
-def active_zone(hass, latitude, longitude, radius=0):
-    """Find the active zone for given latitude, longitude."""
-    return run_callback_threadsafe(
-        hass.loop, async_active_zone, hass, latitude, longitude, radius
-    ).result()
-
-
-@bind_hass
-def async_active_zone(hass, latitude, longitude, radius=0):
-    """Find the active zone for given latitude, longitude.
-
-    This method must be run in the event loop.
-    """
-    # Sort entity IDs so that we are deterministic if equal distance to 2 zones
-    zones = (hass.states.get(entity_id) for entity_id
-             in sorted(hass.states.async_entity_ids(DOMAIN)))
-
-    min_dist = None
-    closest = None
-
-    for zone in zones:
-        if zone.attributes.get(ATTR_PASSIVE):
-            continue
-
-        zone_dist = distance(
-            latitude, longitude,
-            zone.attributes[ATTR_LATITUDE], zone.attributes[ATTR_LONGITUDE])
-
-        within_zone = zone_dist - radius < zone.attributes[ATTR_RADIUS]
-        closer_zone = closest is None or zone_dist < min_dist
-        smaller_zone = (zone_dist == min_dist and
-                        zone.attributes[ATTR_RADIUS] <
-                        closest.attributes[ATTR_RADIUS])
-
-        if within_zone and (closer_zone or smaller_zone):
-            min_dist = zone_dist
-            closest = zone
-
-    return closest
 
 
 def in_zone(zone, latitude, longitude, radius=0) -> bool:
